Added ebuild for glshim (https://github.com/lunixbochs/glshim/)
Also it needs a special version of glu (glues), which makes packaging
a bit tricky because glues is not a drop-in replacement for glu.
So we try to provide a new variant of "media-libs/glu" as
"media-libs/glumesa", which installs into /usr/lib/opengl/xorg-x11
instead of /usr/lib and works with eselect opengl. Then there is
"media-libs/glues" ebuild, which provides glues implementation.
And finally, "media-libs/glshim" provides the glshim opengl
implementation, which only has libGL.so and symlinks to glues
for GLU and to Mesa for libEGL.so, libGLESv1_CM.so, libGLESv2.so

As the end result, "eselect opengl set glshim" should provide
the glshim OpenGL implementation on top of Mesa OpenGL ES.